Respect splitright (#257)
d0210bb5fecfeae58d82a65cfb55f5dffb088052 introduced the behavior to always
move a vertical split to the extreme left, when coc-explorer is on the
right (and vice-versa). This fix the problem of having the coc-explorer
window not fixed on the 'left' or 'right' of vim, but it does not
respect the vim splitright option.

To make the split behavior more predictable, instead of opening the file
from the coc-explorer window, open it from the last window perspective.
This way coc-explorer will respect any split/vsplit options.